+++ src/pages/about/About.test.tsx
@@ -1,0 +1,16 @@
+import React from 'react';
+import { render, screen } from '@testing-library/react';
+import { BrowserRouter as Router } from 'react-router-dom';
+
+import About from './About';
+
+test('is About page exist ', () => {
+  render(<About />);
+  const wraper = screen.getByTestId('about');
+  expect(wraper).toBeInTheDocument();
+});
+test('is About page exist ', () => {
+  render(<About />);
+  const wraper = screen.getByTestId('about');
+  expect(wraper.textContent).toBe('About');
+});
