+++ src/pages/about/About.tsx
@@ -1,7 +1,11 @@
 import React from 'react';
 
 const About = () => {
-  return <div className="container mx-auto my-[50px] px-4">About</div>;
+  return (
+    <div className="container mx-auto my-[50px] px-4" data-testid="about">
+      About
+    </div>
+  );
 };
 
 export default About;
